cleaned up ICICI Dinesh Bank Data for Jan 2022. updated test cases.

+++ src/main/java/com/home/expense/tracker/usercases/reports/BankTransactionDataRow.java
@@ -22,5 +22,5 @@ public interface BankTransactionDataRow {
     int id();
 
     double balance();
-    void balance(Long balance);
+    void balance(double balance);
 }
